do not pass input_config_file twice, allow for docker run without env variables set

+++ entrypoint.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
-cp /ftp-deployment/provide-config.php provide-config.php
-deployment provide-config.php $*
-rm provide-config.php
+set -e
+
+if [ ${INPUT_CONFIG_FILE} ]; then
+  cp /ftp-deployment/provide-config.php provide-config.php
+  deployment provide-config.php ${INPUT_PARAMETERS} $@
+  rm provide-config.php
+else
+  deployment $*
+fi
